Fix AppleScript 'save' Command (#877)
* Fix odd behavior with the save command when using the 'as' parameter.

Previously, the 'save' command only worked for files using the default extension as configured in the app's settings. Apple's default implementation of 'save' includes an 'as' parameter, which would cause an error if used. This commit ensures 'save' works correctly whether 'as' is specified or not.

* Allow 'save' command for TextBundles

* Improve support for 'as' command, cleanup

Avoid various errors when using AppleScript's "save in [file] as [type]" command by checking for extension-type match. Surface more actionable errors in other cases.

* Fix document not reflecting save

* Convention fixes